+++ planets/MainView.hpp
@@ -11,7 +11,7 @@ public:
 	~MainView();
 
 	// Inherited via ViewBase
-	virtual void Update(float fpsFactor) override;
+	virtual void Update(long long fpsFactor) override;
 	virtual void draw(sf::RenderTarget& target, sf::RenderStates states) const override;
 
 private:
