+++ planets/Physics.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "Physics.hpp"
 
 #define _USE_MATH_DEFINES
@@ -20,7 +19,7 @@ Physics::~Physics()
 /// <param name="t"></param>
 /// <param name="g"></param>
 /// <returns></returns>
-float Physics::VerticalThrow(float v0, float t, float g)
+double Physics::VerticalThrow(double v0, double t, double g)
 {
-	return v0 * t - (g / 2) * powf(t, 2.f);
+	return v0 * t - (g / 2) * pow(t, 2);
 }
