+++ planets/ViewBase.hpp
@@ -11,7 +11,7 @@ class ViewBase : public Drawable
     ~ViewBase();
 
     virtual void draw(sf::RenderTarget& target,sf::RenderStates states)const = 0;
-	virtual void Update(float fpsFactor) = 0;
+	virtual void Update(long long fpsFactor) = 0;
 
     private:
